Rival 1 script + free HMs
Gives ability scanner after rival 1 fight. Also changes the HM move scripts to check if a mon can learn the move, not if it knows it. Moves still must be known in order to be used from the party menu. ALSO, THE CALL TO CanMonLearnTMHM in scrcmd.c IS HARD-CODED, SO YOU HAVE TO CHANGE IT IF YOU CHANGE ITEM_TM01'S ID