Mid-battle form changes + Castform split
Allows abilities to cause form changes. Forms are written in the data as separate species. When a form-changing ability's conditions are met in battle_util, it will call TryFormChange, which sets the target species based on the mon's species. If successful, it will call battlescript CastformChange, which swaps the mon's sprite and recalculates its stats according to its new species. Its stats are converted back to the original species when fainting, switching, or ending a battle. (If there are any edge cases to this, the mon will remain in its changed form; fix if that happens.)