Burrows
Allows Dig to be used to dig into "Burrows". Essentially, if the player is standing on a metatile with MB_BURROW_ENTRANCE, using Dig will warp them to the map defined in field_effect.c's sBurrowMapTable.

Each map can only lead to one burrow.

If a map does not have a burrow entry, the burrow entrance instead allows Dig to be used even in maps where it is disallowed. This might be buggy though, I didn't investigate how the Dig exit is set.