TMHM fix and preliminary achievement stuff
Reverts TMHM learnsets to a bitfield that supports 128 TMs.

Also adds the skeleton of the achievement system. When an achievement's criteria are met, VAR_PENDING_ACHIEVEMENTS is increased by one. If VAR_PENDING_ACHIEVEMENTS is non-zero, it is decreased and then the list of achievement criteria are checked. If an achievement is activated, the player will be notified and its flag will be set.

To complete this system, convert the Trainers' Eyes menu into an Achivements menu and convert the pokenav messagebox accordingly.