Match Call overhaul part 2
Added Achievements. Achievement constants are defined in flags.h. There are C and script functions for reading and writing achievements. Each achievement consists of two bits: one to track that the achievement is unlocked, and the other to track that the player has received a popup about it. Currently, they take up 16 bytes in the saveblock (64 achievements).