Shops with limited inventories
Placing a number other than 0 after the inventory of a pokemart command will turn it into a limited shop. The ID corresponds to a set of quantities as set up in new_game.c. When an item's quantity is exhausted in a particular shop, it can no longer be bought. Quantity is printed in the shop menu after price.

TODO: Add a command (special?) for increasing a particular shop's inventory, e.g. if the reward for a quest is restocking a shop.