Preset Mon gifts/"Borrowing"
Adds functionality to the 3 extra parameters of givemon. The sixth parameter is used to affect the data of preset mons defined in src/script_pokemon_util.c. 1 gives the mon specified by the ID given in the 4th parameter to the player ("borrowing"). 2 removes the species of that mon from the player's party ("returning"). Example mon: Gleis's Faceleech. Also adds Faceleech graphics.