Indicators in party menu for start/select functionality
Places indicators at the top of the party menu notifying the player that they can use Select to swap Pokemon, and Start to swap held items. The indicators only appear in "PARTY_MENU_TYPE_FIELD" mode, which I think is the only time they could be used anyway. I didn't do robust testing for the tile allocation, so if something breaks with the graphics, this was probably the cause.

TODO: Actually add functionality for using Start to swap held items.