L/R indicators in battle
Draws sprites that cue the functions of L/R on the battle interface during the player's turn. During wild battles, L can be used to flee and R can be used to throw a Poke Ball. (The indicators are cosmetic and not related to the actual input handling.) Other functions can be added as needed.

Also adds a hypothetical structure for the functions of the tactics menu, which currently do nothing.