Vertical battle move select UI
Rearranges the UI for selecting a move during battle. The moves are displayed in a vertical list, which partially covers the mon sprite. The window on the right displays the move's power, accuracy, category, PP, and type.
The windows used in the Battle Arena are likely incorrect. Also, the palette for the window was changed, so it's possible that some obscure battle window now looks wrong.