Damage numbers in battle
Adds an option to display the amount of damage done on hit in battle, Live A Live style. (This cannot actually be toggled in the option menu yet)
This number does not adjust itself based on max HP or anything like that; it just displays the value of gBattleMoveDamage. This means it will give the player info about the target's HP that they don't normally have.
Future possibilites for polishing:
 - Make the number react to the effect of the move (type effectiveness, critical hits, etc)
 - Reactions could be palette changes or affine anims
 - Possibly redraw the numbers to be more stylistically interesting (currently they're copy/pasted from the normal font)